refactor: schedule create and update method
In create method, to create a schedule  check inf the schedule for the given account exist.
If so delete then and create a new one. In update metho, get the input data and loop over it and get the schedule in the DB and update it.

Warning: find a better way to create schedule